Initial commit - README.md
The first commit in a git repository describes the initial state of that repository before any changes are made - each one that comes after describes a set of changes to the commit that comes before. However, a commit can't be completely empty, so in this case I've added a README file with a couple of lines of explanation of what this project is and what it's for, which will serve as a base for everything we do from now on in.