Set up docker configuration for a new django project.
We're going to use docker to run our project, this is handy because it means we don't have to install everything the application needs on our own computer by hand, docker makes a little virtual computer that has everything we need. This also has the advantage that everyone working on this project has the exact same environment, meaning we don't end up with weird inconsistencies causing us problems. We can also deploy to a server as a docker image too, so we all know our local version is exactly the same as our users will see!

To do this, we add three files.

- `docker-compose.yml`, which tells Docker how to *run* our application,
- this depends on the `Dockerfile`, which tells it what steps it has to go through to *build* the application before running it.
- One of the steps in the dockerfile is to install python and any extra python packages we need, (in our case, Django is one of these). We store these in a file called `requirements.txt`

We're almost ready to run our skeleton application for the first time!