Set up our Django project for the first time!
So, I only had to do one thing here, but quite a lot changed as a consequence.

To set up django, I needed to run a command that sets up a 'skeleton' project for us.
that command is `docker-compose run web django-admin startproject djangoexample .`, which is a little complicated, so i'll break it down:

 - `docker-compose run web` tells docker that we want everything following it to be run on our 'web' service that we defined in docker-compose.yml
 - `django-admin startproject` is the part run inside our 'web' docker container, and tells django to set up a new project.
 - `djangoexample` is the name of our project, so that django knows what to call the folder it creates
 - `.` is 'the current directory, so django knows where to create it.

This automatically a few files for us:

 * `manage.py`: a script for doing various administrative tasks on our site. We shouldn't ever need to modify it, but we will use it from time to time in the terminal.
 * `djangoexample`: This ontains all the python code which configures our project itself. Django creates a 'skeleton' within it for us, but we wil definitely be editing it.
   * `djangoexample/__init__.py`: This is a special blank file generated by django - we won't go into details but it's needed so that Python treats the containing folder (`djangoexample`) as something called a `module` - a collection of related sets of code files that can be treated as a single unit.
   * `djangoexample/asgi.py`: This file is generated by django, and describes how our project will connect to the network and listen for connections from clients. We won't need to edit it and can forget that it's there :-)
   * `djangoexample/settings.py`: This file is generated by django when we initialze the project, with a useful set of defaults - it describes all the settings and configuration of django for our project. We will edit this to add a new application to the project, where we will add our own code.
   * `djangoexample/urls.py`: This describes how the project should respond to requests for different urls. Django generates a 'skeleton' of it for us when we initialize the project. Once we create our first application, it will have a urls file, where we'll create urls for our own pages.
   * `djaongoexample/wsgi.py`: Like `asgi.py`, this file is generated by django and concerns the way our project connects to the internet. We won't need to edit it, and can just forget about it for now!

Once we've done that, we can run `docker-compose up`, go to `http://localhost:8000` in our browser, and see our project already up and running with the default django page!

However, one tiny thing happens when we do this that I needed to address - Django adds a database by default that creates a file called `db.sqlite3` in our project directory. We're not interested in this, and don't want to add it to git, so i just add a line to our .gitignore file so that it's ignored and never gets added to commits.