Add a more complicated template which generates html from data
We add a second page, /fruits which starts to show a little better what the use of something like Django is for creating dynamic, data driven pages.
This page is rendered with the template `fruit_list.html` which renders a list of fruit provided by python. We also add to the urls.py and views.py files as before to wire up this template to a url, as well as adding some sample fruit data to `models.py`, which is loaded in the view and passed through to the template.